[#36] [#37] refactor: move user interfaces into shared package
