+++ backend/user-service/package.json
@@ -32,6 +32,7 @@
         "@repo/eslint-config": "*",
         "@repo/typescript-config": "*",
         "@repo/request-types": "*",
+        "@repo/user-types": "*",
         "@testcontainers/mongodb": "^10.13.1",
         "@types/bcrypt": "^5.0.2",
         "@types/cors": "^2.8.17",
