+++ backend/user-service/src/models/user.model.ts
@@ -1,7 +1,7 @@
+import { Proficiency } from '@repo/user-types/Proficiency'
+import { Role } from '@repo/user-types/Role'
 import { Schema } from 'mongoose'
 import { IUser } from '../types/IUser'
-import { Proficiency } from '../types/Proficiency'
-import { Role } from '../types/Role'
 
 const userSchema = new Schema<IUser>(
     {
