+++ packages/user-types/package.json
@@ -1,0 +1,11 @@
+{
+    "name": "@repo/user-types",
+    "version": "0.0.0",
+    "private": true,
+    "files": [
+        "IAccessTokenPayload.ts",
+        "IUserDto.ts",
+        "Proficiency.ts",
+        "Role.ts"
+    ]
+}
